決済方法のデータ型をINTに修正(Java側で取り扱いやすいように)

+++ webapp/src/main/java/com/s_giken/training/webapp/controller/editor/PaymentMethodEditorSupport.java
@@ -6,21 +6,22 @@ import com.s_giken.training.webapp.model.PaymentMethod;
 
 /**
  * PaymentMethod 列挙型に対する型変換を行うクラス
- * 
- * 主にHTTPリクエスト文字列からPaymentMetho列挙型への変換、またはその逆を行う。
+ *
+ * 主にHTTPリクエスト文字列からPaymentMethod列挙型への変換、またはその逆を行う。
  */
 public class PaymentMethodEditorSupport extends PropertyEditorSupport {
+
     /**
      * 数値の文字列(支払方法コード)から PaymentMethod列挙型へ変換する
-     * 
+     *
      * リクエスト → モデルオブジェクトのプロパティへ変換する際に利用される。
-     * 
+     *
      * @param text 数値の文字列(支払方法コード)
      */
     @Override
     public void setAsText(String text) {
         try {
-            setValue(PaymentMethod.fromCode(Byte.valueOf(text)));
+            setValue(PaymentMethod.fromCode(Integer.valueOf(text)));
         } catch (IllegalArgumentException e) {
             setValue(PaymentMethod.UNKNOWN);
         }
@@ -28,7 +29,7 @@ public class PaymentMethodEditorSupport extends PropertyEditorSupport {
 
     /**
      * PaymentMethod列挙型オブジェクトから数値の文字列(支払方法コード)へ変換する
-     * 
+     *
      * @return 数値の文字列(支払方法コード)
      */
     @Override
